requirements.txt
volcenginesdkarkruntime
openai
edge-tts
mutagen
playwright
playwright-stealth
python-dotenv
certifi
fastapi
uvicorn[standard]
sqlalchemy
